여태한거..

+++ app/src/main/java/com/mobile/upway/dto/Cheese.java
@@ -5,18 +5,18 @@ import java.io.Serializable;
 public class Cheese implements Serializable {
 
     public String name;
-    public int kcal;
+    public double kcal;
     public String imgUrl;
 
     public Cheese() {
     }
 
-    public Cheese(String name, int kcal) {
+    public Cheese(String name, double kcal) {
         this.name = name;
         this.kcal = kcal;
     }
 
-    public Cheese(String name, int kcal, String imgUrl) {
+    public Cheese(String name, double kcal, String imgUrl) {
         this.name = name;
         this.kcal = kcal;
         this.imgUrl = imgUrl;
@@ -30,11 +30,11 @@ public class Cheese implements Serializable {
         this.name = name;
     }
 
-    public int getKcal() {
+    public double getKcal() {
         return kcal;
     }
 
-    public void setKcal(int kcal) {
+    public void setKcal(double kcal) {
         this.kcal = kcal;
     }
