+++ app/src/main/java/com/mobile/upway/dto/Sauce.java
@@ -5,12 +5,12 @@ import java.io.Serializable;
 public class Sauce implements Serializable {
 
     public String name;
-    public int kcal;
+    public double kcal;
     public String imgUrl;
 
     public Sauce(){}
 
-    public Sauce(String name, int kcal, String imgUrl) {
+    public Sauce(String name, double kcal, String imgUrl) {
         this.name = name;
         this.kcal = kcal;
         this.imgUrl = imgUrl;
@@ -24,11 +24,11 @@ public class Sauce implements Serializable {
         this.name = name;
     }
 
-    public int getKcal(){
+    public double getKcal(){
         return kcal;
     }
 
-    public void setKcal(int kcal){
+    public void setKcal(double kcal){
         this.kcal = kcal;
     }
